Use forward-compatible `FrontendMetadata` creation
Summary:
The `FrontendMetadata` type no longer has a `_replace` method, as of:
<https://github.com/tensorflow/tensorboard/pull/2606>

This patch updates the TF Lite plugin to use a form that is compatible
with both the old and new versions. Once the WORKSPACE dependency on
TensorBoard is updated to the latest version, we can make this more
idiomatic by removing the fields with default values, which will no
longer need to be set explicitly. (See built-in TensorBoard plugins for
examples.)

Test Plan:
All tests pass (`bazel test //...`).

wchargin-branch: frontendmetadata-struct-compat